Redid data fetching format, clean up
The data stored in python uses date, disease, and region as an
multi-index. Before, the index was jsonified as a tuple. Now, it is
jsonified as part of the dict for each data point. Makes front end code
make more sense and less reliant on random indexing
fixed the typo averanged to averaged
moved the bubble static css to the css file

also added some stuff in prep for case/death switch